requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
pandas>=2.1.0,<2.3
openpyxl>=3.1.0
python-multipart>=0.0.6
anthropic>=0.18.0
openai>=1.12.0
requests>=2.31.0
reportlab>=4.0.0
arabic_reshaper>=3.0.0
python-bidi>=0.4.2
sqlalchemy>=2.0.0
pyjwt[crypto]>=2.8.0
bcrypt>=4.1.0
psycopg2-binary>=2.9.9
google-auth>=2.25.0
pyotp>=2.9.0
qrcode[pil]>=7.4.0
redis>=5.0.0
sentry-sdk[fastapi]>=1.40.0
python-json-logger>=2.0.7

# Optional: S3 storage backend (only needed if STORAGE_BACKEND=s3)
# boto3>=1.34.0

# Optional: Stripe payment backend (only needed if PAYMENT_BACKEND=stripe)
# stripe>=7.0.0

# Test dependencies
# pytest>=7.4.0
# httpx>=0.25.0
# pytest-asyncio>=0.23.0
